requirements.txt
hewwwo
discord
python-dotenv
